Move Registry.scala
